Hash utility moved to loader-utils
