+++ modules/i3s/src/lib/parsers/parse-slpk/slpk-archieve.ts
@@ -1,34 +1,7 @@
 import md5 from 'md5';
 import {FileProvider, parseZipLocalFileHeader} from '@loaders.gl/zip';
 import {GZipCompression} from '@loaders.gl/compression';
-
-/** Element of hash array */
-export type HashElement = {
-  /** File name hash */
-  hash: Buffer;
-  /** File offset in the archive */
-  offset: bigint;
-};
-
-/**
- * Comparing md5 hashes according to https://github.com/Esri/i3s-spec/blob/master/docs/2.0/slpk_hashtable.pcsl.md step 5
- * @param hash1 hash to compare
- * @param hash2 hash to compare
- * @returns -1 if hash1 < hash2, 0 of hash1 === hash2, 1 if hash1 > hash2
- */
-export const compareHashes = (hash1: Buffer, hash2: Buffer): number => {
-  const h1 = new BigUint64Array(hash1.buffer, hash1.byteOffset, 2);
-  const h2 = new BigUint64Array(hash2.buffer, hash2.byteOffset, 2);
-
-  const diff = h1[0] === h2[0] ? h1[1] - h2[1] : h1[0] - h2[0];
-
-  if (diff < 0n) {
-    return -1;
-  } else if (diff === 0n) {
-    return 0;
-  }
-  return 1;
-};
+import {HashElement, findBin} from '@loaders.gl/loader-utils';
 
 /** Description of real paths for different file types */
 const PATH_DESCRIPTIONS: {test: RegExp; extensions: string[]}[] = [
@@ -76,29 +49,6 @@ export class SLPKArchive {
     this.slpkArchive = slpkArchive;
     this.hashArray = hashFile;
   }
-
-  /**
-   * Binary search in the hash info
-   * @param hashToSearch hash that we need to find
-   * @returns required hash element or undefined if not found
-   */
-  private findBin = (hashToSearch: Buffer): HashElement | undefined => {
-    let lowerBorder = 0;
-    let upperBorder = this.hashArray.length;
-
-    while (upperBorder - lowerBorder > 1) {
-      const middle = lowerBorder + Math.floor((upperBorder - lowerBorder) / 2);
-      const value = compareHashes(this.hashArray[middle].hash, hashToSearch);
-      if (value === 0) {
-        return this.hashArray[middle];
-      } else if (value < 0) {
-        lowerBorder = middle;
-      } else {
-        upperBorder = middle;
-      }
-    }
-    return undefined;
-  };
 
   /**
    * Returns file with the given path from slpk archive
@@ -167,7 +117,7 @@ export class SLPKArchive {
    */
   private async getFileBytes(path: string): Promise<ArrayBuffer | undefined> {
     const nameHash = Buffer.from(md5(path), 'hex');
-    const fileInfo = this.findBin(nameHash); // implement binary search
+    const fileInfo = findBin(nameHash, this.hashArray); // implement binary search
     if (!fileInfo) {
       return undefined;
     }
@@ -183,5 +133,8 @@ export class SLPKArchive {
     );
 
     return compressedFile;
+  }
+  findBin(nameHash: Buffer) {
+    throw new Error('Method not implemented.');
   }
 }
