+++ modules/loader-utils/src/index.ts
@@ -117,3 +117,5 @@ export type {WritableFile} from './lib/filesystems/writable-file';
 export {makeWritableFile} from './lib/filesystems/writable-file';
 
 export {default as _NodeFileSystem} from './lib/filesystems/node-filesystem';
+
+export {HashElement, compareHashes, parseHashFile, findBin} from './hash-file-utility';
